+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Version 1.1.2
 - Fix for package.json that has a missing name
 - Fix to all script running on a vanilla web project
+- Fix for handling authentication issue when running npm outdated
 
 ### Version 1.1.1
 - Remove cordova feature now works on Windows